requirements.txt
opencv-python>=4.5.0
numpy>=1.20.0
flask>=2.0.0
transformers>=4.15.0
torch>=1.10.0
pillow>=8.0.0
waitress>=2.0.0